+++ target-spec/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [2.0.0] - 2023-06-19
 
 ### Added
 
@@ -246,6 +246,7 @@ This was mistakenly published and was yanked.
 
 - Initial release.
 
+[2.0.0]: https://github.com/guppy-rs/guppy/releases/tag/target-spec-2.0.0
 [1.4.0]: https://github.com/guppy-rs/guppy/releases/tag/target-spec-1.4.0
 [1.3.1]: https://github.com/guppy-rs/guppy/releases/tag/target-spec-1.3.1
 [1.3.0]: https://github.com/guppy-rs/guppy/releases/tag/target-spec-1.3.0